// wdip-fe/src/service/wdip-be/models/Motions.ts
// tslint:disable
/**
 * WDIP
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.0.1
 * Contact: markus@silberstein.nu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
import {
    Motion,
    MotionFromJSON,
    MotionToJSON,
} from './';

/**
 * 
 * @export
 * @interface Motions
 */
export interface Motions {
    /**
     * 
     * @type {Array<Motion>}
     * @memberof Motions
     */
    results?: Array<Motion>;
    /**
     * 
     * @type {number}
     * @memberof Motions
     */
    total?: number;
    /**
     * 
     * @type {number}
     * @memberof Motions
     */
    startResult?: number;
    /**
     * 
     * @type {number}
     * @memberof Motions
     */
    endResult?: number;
}

export function MotionsFromJSON(json: any): Motions {
    return {
        'results': !exists(json, 'results') ? undefined : (json['results'] as Array<any>).map(MotionFromJSON),
        'total': !exists(json, 'total') ? undefined : json['total'],
        'startResult': !exists(json, 'startResult') ? undefined : json['startResult'],
        'endResult': !exists(json, 'endResult') ? undefined : json['endResult'],
    };
}

export function MotionsToJSON(value?: Motions): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'results': (value.results as Array<any>).map(MotionToJSON),
        'total': value.total,
        'startResult': value.startResult,
        'endResult': value.endResult,
    };
}