// wdip-fe/src/service/wdip-be/models/Motion.ts
// tslint:disable
/**
 * WDIP
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.0.1
 * Contact: markus@silberstein.nu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
import {
    Proposal,
    ProposalFromJSON,
    ProposalToJSON,
    Stakeholder,
    StakeholderFromJSON,
    StakeholderToJSON,
} from './';

/**
 * 
 * @export
 * @interface Motion
 */
export interface Motion {
    /**
     * 
     * @type {string}
     * @memberof Motion
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Motion
     */
    documentStatus: string;
    /**
     * 
     * @type {string}
     * @memberof Motion
     */
    documentType: string;
    /**
     * 
     * @type {Array<Proposal>}
     * @memberof Motion
     */
    proposals: Array<Proposal>;
    /**
     * 
     * @type {Array<Stakeholder>}
     * @memberof Motion
     */
    stakeholders: Array<Stakeholder>;
}

export function MotionFromJSON(json: any): Motion {
    return {
        'id': json['id'],
        'documentStatus': json['documentStatus'],
        'documentType': json['documentType'],
        'proposals': (json['proposals'] as Array<any>).map(ProposalFromJSON),
        'stakeholders': (json['stakeholders'] as Array<any>).map(StakeholderFromJSON),
    };
}

export function MotionToJSON(value?: Motion): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'id': value.id,
        'documentStatus': value.documentStatus,
        'documentType': value.documentType,
        'proposals': (value.proposals as Array<any>).map(ProposalToJSON),
        'stakeholders': (value.stakeholders as Array<any>).map(StakeholderToJSON),
    };
}